refactor(js): clean up modal_handler.js
Since there is a single modal in miniflux, namely the keyboard help, there is
no need to have a full-fledged focus stealer where an ad-hoc one would be
enough. This simplifies the code and makes it way more understandable.

We might want to move to the <dialog> tag at some point in the future.